+++ packages/manager/src/hooks/useInterfaceDataForLinode.ts
@@ -1,9 +1,4 @@
-import {
-  useLinodeConfigQuery,
-  useLinodeFirewallsQuery,
-  useLinodeInterfaceFirewallsQuery,
-  useLinodeInterfaceQuery,
-} from '@linode/queries';
+import { useLinodeConfigQuery, useLinodeInterfaceQuery } from '@linode/queries';
 
 /**
  * In some files, We need to handle support for both legacy interfaces and Linode interfaces.
@@ -11,38 +6,13 @@ import {
  * To fetch data, depending if the Linode with the given ID is a Linode Interface, we use Linode Interface related queries.
  * Otherwise, we use config profile interface related queries.
  */
-export const useInterfaceAndFirewallDataForLinode = (inputs: {
+export const useInterfaceDataForLinode = (inputs: {
   configId: null | number;
   interfaceId: number;
   isLinodeInterface: boolean;
   linodeId: number;
 }) => {
   const { configId, interfaceId, isLinodeInterface, linodeId } = inputs;
-
-  // query to fetch firewalls if this Linode uses config profile interfaces
-  const {
-    data: attachedFirewallsConfig,
-    error: firewallsErrorConfig,
-    isLoading: firewallsLoadingConfig,
-  } = useLinodeFirewallsQuery(linodeId, !isLinodeInterface);
-
-  // query to fetch firewalls for a Linode Interface (firewalls are attached at the interface level)
-  const {
-    data: attachedFirewallsLinodeInterface,
-    error: firewallsErrorLinodeInterface,
-    isLoading: firewallsLoadingLinodeInterface,
-  } = useLinodeInterfaceFirewallsQuery(
-    linodeId,
-    interfaceId,
-    isLinodeInterface
-  );
-
-  // Depending on which query was fired to fetch firewalls, return appropriate data
-  const attachedFirewalls =
-    attachedFirewallsConfig ?? attachedFirewallsLinodeInterface;
-  const firewallsError = firewallsErrorConfig ?? firewallsErrorLinodeInterface;
-  const firewallsLoading =
-    firewallsLoadingConfig || firewallsLoadingLinodeInterface;
 
   const {
     data: linodeInterface,
@@ -69,11 +39,6 @@ export const useInterfaceAndFirewallDataForLinode = (inputs: {
   const interfaceLoading = linodeInterfaceLoading ?? configLoading;
 
   return {
-    firewallsInfo: {
-      attachedFirewalls,
-      firewallsError,
-      firewallsLoading,
-    },
     interfacesInfo: {
       config, // undefined if this Linode is using Linode Interfaces. Used to determine an unrecommended configuration
       configInterface, // undefined if this Linode is using Linode Interfaces
